+++ types/index.d.ts
@@ -1,10 +1,10 @@
 /**
  * @name Cause & Effect
- * @version 1.3.0
+ * @version 1.3.1
  * @author Esther Brunner
  */
 export { CircularDependencyError, type Guard, InvalidCallbackError, InvalidSignalValueError, NullishSignalValueError, ReadonlySignalError, RequiredOwnerError, UnsetSignalValueError, } from './src/errors';
-export { batch, type Cleanup, type ComputedOptions, createScope, DEFAULT_EQUALITY, DEEP_EQUALITY, type EffectCallback, type MaybeCleanup, type MemoCallback, type Signal, type SignalOptions, SKIP_EQUALITY, type TaskCallback, unown, untrack, } from './src/graph';
+export { batch, type Cleanup, type ComputedOptions, createScope, DEFAULT_EQUALITY, DEEP_EQUALITY, type EffectCallback, type MaybeCleanup, type MemoCallback, type ScopeOptions, type Signal, type SignalOptions, SKIP_EQUALITY, type TaskCallback, unown, untrack, } from './src/graph';
 export { type Collection, type CollectionCallback, type CollectionChanges, type CollectionOptions, createCollection, type DeriveCollectionCallback, isCollection, } from './src/nodes/collection';
 export { createEffect, type MatchHandlers, type SingleMatchHandlers, type MaybePromise, match, } from './src/nodes/effect';
 export { createList, isList, type KeyConfig, type List, type ListOptions, } from './src/nodes/list';